refactor: use functional pattern instead of factory/adapter
