+++ governance/xc_admin/packages/xc_admin_common/src/index.ts
@@ -15,9 +15,6 @@ export * from "./deterministic_stake_accounts";
 export * from "./price_store";
 export { default as lazerIdl } from "./multisig_transaction/idl/lazer.json";
 
-// Export program-related modules
-export * from "./programs/program_adapter";
+// Export program-related modules - new functional approach
 export * from "./programs/types";
-export * from "./programs/adapter_factory";
-export * from "./programs/core/core_adapter";
-export * from "./programs/lazer/lazer_adapter";
+export * from "./programs/program_registry";
